backup_20250802_192644/requirements.txt
streamlit
pdfplumber
pyperclip
openai
anthropic
python-dotenv

# Performance and optimization
tiktoken>=0.5.0
cachetools>=5.0.0

# Error handling and monitoring
requests>=2.28.0
